some changes in updatebook section

+++ frontend/src/components/AllBooks.jsx
@@ -1,12 +1,14 @@
 import React, { useEffect, useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom';
 import axios from "axios";
+import Swal from 'sweetalert2'
 
 
 const AllBooks = () => {
 
     const [books , setBooks] = useState([]);
     const navigate = useNavigate();
+    const [reloadData, setReloadData] = useState(true);
 
     // const [deletebook , setDeletebook] =useState();
 
@@ -17,11 +19,33 @@ const AllBooks = () => {
        console.log(data);
     }
 
-    const deleteBook = async (id)=>{
+    const deleteBook = async (id,event)=>{
+        event.preventDefault();
       // const res = await fetch(`http://localhost:4000/books/deletebook/${id}`);
+      
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      })
+      .then(async(result) => {
+        if (result.isConfirmed) {
+          
+           await axios.delete(`http://localhost:4000/books/deletebook/${id}`);
+            setReloadData((prev) => !prev);
+          
 
-      axios.delete(`http://localhost:4000/books/deletebook/${id}`);  
-      window.location.reload();
+        //  axios.delete(`http://localhost:4000/books/deletebook/${id}`);  
+        //  alert("deleted!!");
+        //  window.location.reload();
+        }
+      });
+
+     
 
     }
 
@@ -29,7 +53,7 @@ const AllBooks = () => {
 
     useEffect(()=>{
         fetchBooks();
-    },[])
+    },[reloadData])
 
   return (
     <div>
@@ -44,7 +68,7 @@ const AllBooks = () => {
            
            <div className={`gap-3 mt-3 ${e.name === "error in backend" ? "hidden":"flex"} `} >
            <button className='border-2 border-green-500 p-2 bg-green-500 text-white' > <Link to={`/updatebook/${e.book_id}` }>Update</Link></button>
-           <button className='border-2 border-red-500 p-2 bg-red-500 text-white' onClick={()=>{deleteBook(e.book_id)}} > Delete</button>
+           <button className='border-2 border-red-500 p-2 bg-red-500 text-white' onClick={(event)=>{deleteBook(e.book_id,event)}} > Delete</button>
            </div>
         
         </div>
@@ -53,7 +77,7 @@ const AllBooks = () => {
 
       </div>
 
-    <div className='flex justify-center my-40'> <button className='border-2 border-blue-500 bg-blue-500 px-2 py-3 text-white'> <Link to={"/addbook"}>Add New Book </Link> </button> </div>
+    <div className='flex justify-center my-40'> <button className='border-2 border-blue-500 bg-blue-500 hover:bg-blue-600 px-2 py-3 text-white'> <Link to={"/addbook"}>Add New Book </Link> </button> </div>
 
     </div>
   )
